requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1
PyMySQL==1.1.1
redis==5.0.7
python-dotenv==1.0.1
pydantic==2.8.2
pandas==2.2.2
numpy==2.0.1
pyspark==3.5.2
